Add Role type to ARG_TYPE
